separate backend updates for shared and static libs into separate PR
